task(refactor): refactor codebase
- add modules for different sections
- change imports to es6 standards
- transfer verification endpoint

[Finishes]
