+++ src/modules/bot/helpers/validation.js
@@ -1,0 +1,8 @@
+function validate(req, res) {
+  if (req.query['hub.verify_token'] === 'chirchir') {
+    res.send(req.query['hub.challenge']);
+  }
+  res.send('wrong token');
+}
+
+export default validate;
