+++ src/modules/index.js
@@ -1,0 +1,7 @@
+import botRouter from './bot';
+
+const routes = (app) => {
+  app.use(botRouter);
+};
+
+export default routes;
